\name{ChangepointTesting-package}
\alias{ChangepointTesting-package}
\alias{ChangepointTesting}
\docType{package}
\title{ Change Point Estimation for Clustered Signals }
\description{ A multiple testing procedure for clustered alternative 
              hypotheses. It is assumed that the p-values under the 
              null hypotheses follow U(0,1) and that the distributions
              of p-values from the alternative hypotheses are
              stochastically smaller than U(0,1).
              By aggregating information, this method is more
              sensitive to detecting signals of low magnitude than
              standard methods. Additionally,
              sporadic small p-values appearing within a
              null hypotheses sequence are avoided 
              by averaging on the neighboring p-values. }
\details{
\tabular{ll}{
Package: \tab ChangepointTesting\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-02-02\cr
License: GPL-2\cr
}
}
\author{
Hongyuan Cao, Wei Biao Wu, and Shannon T.  Holloway
Maintainer: Shannon T. Holloway <sthollow@ncsu.edu>
}
\references{
Cao, H. and Wu, W. B. (2015)
Changepoint estimation: Another look at multiple testing problems.
Biometrika, 102, 974--980.
}
\keyword{ package }

